Add fix for gen_v simm5 extraction
Compilation of Renode yielded error for this portion of gen_v.

```
tlib/arch/riscv/translate.c:1899:13: error: a label can only be part
of a statement and a declaration is not a statement
```

This change just separates the declaration and the initializatoin into
separate lines, which then allows for compilation to proceed
without errors.

Added a # Please enter the commit message for your changes. Lines starting