Add script and opcodes file for parsing
These are not meant to live here, but are relevant to the
subsequent commits which utilize te results of running
parse_opcodes.py on the opcodes-rvv file.

The opcodes-rvv file is from the RiscV rvv-opcodes repo, and is
built with regex based parsing in mind.

We use the parse_opcodes.py in to create masks for subsequent commits,
mainly with the `--base hex` option.