Add basis for counting operations
The rvv_opcode_map.h file contains a formatted output of from running
parse_opcodes.py with the `--base hex` option.

Added a flag for turning on opcode counting in the cpu.h CPUState
struct, as well as an array to cpu.h for counting the rvv instructions.

The translate.c file has one check to see if the instructions are
flagged to be counted, to minimize the effect on normal runtime.

The counter then (essentially) loops through the opcodes to search for a
match, incrementing the count in the array (see `rvv_inst_counter.c`
for more details).