Expose methods to the Renode monitor
Added functions that permit toggling the enable bit,
as well as checking for specific vector instruction counts.

These are to be utilized on the C# descriptions (e.g. BaseRiscV.cs),
where one can utilize a dictionary to quickly check instruction counts
after running an elf, and save out to a csv.